+++ pkg/disttask/framework/storage/task_table.go
@@ -16,7 +16,6 @@ package storage
 
 import (
 	"context"
-	"encoding/json"
 	"strconv"
 	"strings"
 	"sync/atomic"
@@ -824,36 +823,4 @@ func (mgr *TaskManager) AdjustTaskOverflowConcurrency(ctx context.Context, se se
 	sql := "update mysql.tidb_global_task set concurrency = %? where concurrency > %?;"
 	_, err = sqlexec.ExecSQL(ctx, se.GetSQLExecutor(), sql, cpuCount, cpuCount)
 	return err
-}
-
-// ModifyTaskByID modifies the task by the task ID.
-func (mgr *TaskManager) ModifyTaskByID(ctx context.Context, taskID int64, param *proto.ModifyParam) error {
-	if param.PrevState != proto.TaskStatePending &&
-		param.PrevState != proto.TaskStateRunning &&
-		param.PrevState != proto.TaskStatePaused {
-		return ErrTaskStateNotAllow
-	}
-	bytes, err := json.Marshal(param)
-	if err != nil {
-		return errors.Trace(err)
-	}
-	return mgr.WithNewSession(func(se sessionctx.Context) error {
-		_, err = mgr.getTaskBaseByID(ctx, se.GetSQLExecutor(), taskID)
-		if err != nil {
-			return err
-		}
-		_, err = sqlexec.ExecSQL(ctx, se.GetSQLExecutor(), `
-			update mysql.tidb_global_task
-			set state = %?, modify_params = %?, state_update_time = CURRENT_TIMESTAMP()
-			where id = %? and state = %?`,
-			proto.TaskStateModifying, json.RawMessage(bytes), taskID, param.PrevState,
-		)
-		if err != nil {
-			return err
-		}
-		if se.GetSessionVars().StmtCtx.AffectedRows() == 0 {
-			return ErrTaskStateNotAllow
-		}
-		return nil
-	})
 }
